% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_controls_maplibre.R
\name{add_fullscreen}
\alias{add_fullscreen}
\title{Add MapLibre fullscreen control}
\usage{
add_fullscreen(map, position = "topright", ...)
}
\arguments{
\item{map}{A maplamina widget created by \code{\link[=maplamina]{maplamina()}}.}

\item{position}{Control position: \code{"topleft"}, \code{"topright"}, \code{"bottomleft"}, \code{"bottomright"}.}

\item{...}{Named options passed through to the underlying MapLibre control.}
}
\value{
The modified map widget.
}
\description{
Add MapLibre fullscreen control
}
\examples{
maplamina() |>
  add_fullscreen()
}
