% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_panel.R
\name{sections}
\alias{sections}
\title{Create a list of panel sections}
\usage{
sections(...)
}
\arguments{
\item{...}{One or more \code{section()} objects, character ids, or lists of sections.}
}
\description{
Create a list of panel sections
}
