% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_bin}
\alias{color_bin}
\title{Binned (equal-interval or manual breaks) color scale specification (resolved during compilation).}
\usage{
color_bin(
  expr,
  palette = NULL,
  bins = 5,
  domain = NULL,
  na_color = "#00000000",
  reverse = FALSE,
  clamp = TRUE
)
}
\description{
Binned (equal-interval or manual breaks) color scale specification (resolved during compilation).
}
