% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_factor}
\alias{color_factor}
\title{Categorical color scale specification (resolved during compilation).}
\usage{
color_factor(
  expr,
  palette = NULL,
  domain = NULL,
  na_color = "#00000000",
  reverse = FALSE
)
}
\description{
Categorical color scale specification (resolved during compilation).
}
