% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_numeric}
\alias{color_numeric}
\title{Continuous color scale specification (resolved during compilation).}
\usage{
color_numeric(
  expr,
  palette = NULL,
  domain = NULL,
  steps = 256L,
  na_color = "#00000000",
  reverse = FALSE,
  clamp = TRUE
)
}
\description{
Continuous color scale specification (resolved during compilation).
}
