% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_quantile}
\alias{color_quantile}
\title{Quantile color scale specification (resolved during compilation).}
\usage{
color_quantile(
  expr,
  palette = NULL,
  n = 5,
  domain = NULL,
  na_color = "#00000000",
  reverse = FALSE,
  clamp = TRUE
)
}
\description{
Quantile color scale specification (resolved during compilation).
}
